stepSize -> int tidyup

+++ scripts/Game/RECOM/DTOs/Map/MapTopographyScanner/RECOM_MapTopographyEntityDto.c
@@ -1,7 +1,7 @@
 class RECOM_MapTopographyEntityDto : JsonApiStruct {
 
 	// META DATA
-	float stepSize;
+	int stepSize;
 	int scanIterationsX;
 	int scanIterationsZ;
 	float oceanHeight;
